Introduce MVC pattern into pin out tab component
Next part of refactoring application main components
Use MVC architectural pattern to organize
code and folder structure in gear component
- Move component management logic from PinOutTab to newly
  created PinOutTabController that derrives from
  ITabController
- Rename subtab DataEdit into EntryDataTab and
  move it from gear_tab to separate subtabs folder
- Organize application logic
CAUTION - the MVC pattern was applied only on the surface
level - it needs refinement deeper in the subtabs structure
