+++ Tabs/pin_out/subtabs/entry_data/entry_data_tab.py
@@ -1,27 +1,18 @@
-from functools import partial
 from math import pi
 from typing import Dict, Optional, Union, Tuple
 
 from PySide2.QtCore import Signal
-from PySide2.QtWidgets import QWidget, QGridLayout, QVBoxLayout, QPushButton, QStackedLayout, QCheckBox
+from PySide2.QtWidgets import QWidget, QGridLayout, QPushButton
 
 from tabs.common.widgets.abstract_tab import AbstractTab
-from common.common_widgets import DoubleSpinBox, QLabelD, IntSpinBox, ResponsiveContainer
-from common.utils import open_pdf
-from ..common.widgets.tolerance_widgets import ToleranceEdit
-from ..common.widgets.charts import ResultsTab
-from .Mech_wyj_tuleje.popups import SupportWin
-from .Mech_wyj_tuleje.tuleje_obl import obliczenia_mech_wyjsciowy
-from .Mech_wyj_tuleje.utils import sprawdz_przecinanie_otworow
-from .Mech_wyj_tuleje.widgets import ResultsFrame, MaterialsFrame
-#TODO: mam przesunięte do tyłu w poziomie punkty wykresów. Pawel też.
-# moze jakos usuwac dane z wykresow jak sa bledy?
-# po trzy razy wysylam dane w niektorych sytuacjach, np jak sie zaznaczy zeby uzywac tego. Może to jest pętla.
-#TODO: moze zrobic w animacji painter.save() i .restore() zamiast obrotów -self.kat_dorotacji
-#TODO: ukrywanie label e2 teraz jak jest QSCrollArea nie dziala, bo to sie dzieje na zewnatrz, pokazanie wszystkich widzetow
+from common.common_widgets import DoubleSpinBox, QLabelD, IntSpinBox
 
+from .popups import SupportWin
+from ...model.tuleje_obl import obliczenia_mech_wyjsciowy
+from .utils import sprawdz_przecinanie_otworow
+from .widgets import ResultsFrame, MaterialsFrame
 
-class DataEdit(QWidget):
+class EntryDataTab(QWidget):
     chartDataUpdated = Signal(dict)
     animDataUpdated = Signal(dict)
     errorsUpdated = Signal(dict)
@@ -315,215 +306,3 @@ class DataEdit(QWidget):
     def toleranceUpdate(self, tol_data: Optional[Dict[str, Union[float, Tuple[float, float]]]]) -> None:
         self.tol_data = tol_data
         self.recalculate()
-
-
-class PinOutTab(AbstractTab):
-    thisEnabled = Signal(bool)
-    def __init__(self, parent: QWidget) -> None:
-        super().__init__(parent)
-
-        layout = QVBoxLayout()
-        button_layout = QGridLayout()
-        stacklayout = QStackedLayout()
-        layout.addLayout(button_layout)
-        layout.addLayout(stacklayout)
-        self.use_this_check = QCheckBox(text="Używaj tego mechanizmu wyjściowego")
-        self.use_this_check.stateChanged.connect(self.useThisChanged)
-        button_layout.addWidget(self.use_this_check, 1, 0, 1, 2)
-
-        help_pdf_button = QPushButton("Pomoc")
-        button_layout.addWidget(help_pdf_button, 1, 2)
-        help_pdf_button.clicked.connect(lambda: open_pdf("help//mechanizmy-sworzniowe-help-1.pdf"))
-
-        self.data = DataEdit(self)
-        self.data.shouldSendData.connect(self.sendData)
-        self.wykresy = ResultsTab(self, "numer sworznia", {
-            "sily": {"repr_name": "Siły", "chart_title": "Siły na sworzniach", "y_axis_title": "Siła [N]"},
-            "naciski": {"repr_name": "Naciski", "chart_title": "Naciski powierzchniowe na sworzniach", "y_axis_title": "Wartość Nacisku [MPa]"},
-            "straty": {"repr_name": "Straty", "chart_title": "Straty mocy na sworzniach", "y_axis_title": "Straty mocy [W]"},
-            "luzy": {"repr_name": "Luzy", "chart_title": "Luzy na sworzniach", "y_axis_title": "Luz [mm]"},
-        })
-        self.tol_edit = ToleranceEdit(self, (
-            ("T_o", "T<sub>o</sub>", "Tolerancja wykonania promieni otworów w kole cykloidalnym"),
-            ("T_t", "T<sub>t</sub>", "Tolerancja wykonania promieni tuleji"),
-            ("T_s", "T<sub>s</sub>", "Tolerancja wykonania promieni sworzni"),
-            ("T_Rk", "T<sub>Rk</sub>", "Tolerancja wykonania promienia rozmieszczenia otworów w kole cykloidalnym"),
-            ("T_Rt", "T<sub>Rt</sub>", "Tolerancja wykonania promienia rozmieszczenia tulei w elemencie wyjściowym"),
-            ("T_fi_k", "T<sub>\u03c6k</sub>", "Tolerancja wykonania kątowego rozmieszczenia otworów w kole cykloidalnym"),
-            ("T_fi_t", "T<sub>\u03c6t</sub>", "Tolerancja wykonania kątowego rozmieszczenia tulei w elemencie wyjściowym"),
-            ("T_e", "T<sub>e</sub>", "Tolerancja wykonania mimośrodu"),
-        ))
-        
-        self.data.chartDataUpdated.connect(self.wykresy.updateResults)
-        self.tol_edit.toleranceDataUpdated.connect(self.data.toleranceUpdate)
-
-        scrollable_tab = ResponsiveContainer(self, self.data, self.data.setupSmallLayout, self.data.setupLayout, 620, 1200)
-        tab_titles = ["Wprowadzanie Danych", "Wykresy", "Tolerancje"]
-        stacked_widgets = [scrollable_tab, self.wykresy, self.tol_edit]
-
-        for index, (title, widget) in enumerate(zip(tab_titles, stacked_widgets)):
-            button = QPushButton(title)
-            button_layout.addWidget(button, 0, index)
-            stacklayout.addWidget(widget)
-            button.pressed.connect(partial(stacklayout.setCurrentIndex, index))
-
-        self.setLayout(layout)
-        self.data.setEnabled(False)
-        self.tol_edit.setEnabled(False)
-    
-    def useThisChanged(self, state: bool) -> None:
-        self.data.setEnabled(state)
-        self.tol_edit.setEnabled(state)
-        if state:
-            self.thisEnabled.emit(True)
-            self.data.module_enabled = True
-            self.data.recalculate()
-        else:
-            self.thisEnabled.emit(False)
-            self.data.module_enabled = False
-            self.data.animDataUpdated.emit({"PinOutTab": False})
-
-    def useOtherChanged(self, state: bool) -> None:
-        self.use_this_check.setEnabled(not state)
-
-    def sendData(self) -> None:
-        M_k = self.data.zew_dane["M_wyj"] / self.data.zew_dane["K"]
-        R_wt = self.data.input_dane["R_wt"]
-        
-        self.dataChanged.emit({"PinOutTab": {
-            "F_wmr": 1000 * 4 * M_k / (pi * R_wt),
-            "r_mr": pi * R_wt / 4,
-        }})
-    
-    def receiveData(self, new_data: Dict[str, Optional[Dict[str, Union[int, float]]]]) -> None:
-        wanted_data = None
-        if new_data.get("GearTab") is not None:
-            wanted_data = new_data.get("GearTab")
-        elif new_data.get("base") is not None:
-            wanted_data = new_data.get("base")
-        else:
-            return
-        
-        for key in wanted_data:
-            if self.data.zew_dane.get(key) is not None:
-                self.data.zew_dane[key] = wanted_data[key]
-
-        # if wanted_data.get("K") == 2:
-        #     self.data.label_e2.show()
-        #     self.data.input_widgets["e2"].show()
-        # elif wanted_data.get("K") == 1:
-        #     self.data.label_e2.hide()
-        #     self.data.input_widgets["e2"].hide()
-        if self.data.accept_button.isEnabled():
-            self.data.sendAnimationUpdates()
-        else:
-            self.data.recalculate()
-        
-        self.sendData()
-
-    def saveData(self) -> Dict:
-        self.data.recalculate()
-        return {
-            "input_dane": self.data.input_dane,
-            "zew_dane": self.data.zew_dane,
-            "material_data": self.data.material_frame.getData(),
-            "tolerancje": self.tol_edit.tolerancje,
-            "tol_mode": self.tol_edit.mode,
-            "use_tol": self.tol_edit.check.isChecked()
-        }
-
-    def loadData(self, new_data: Dict) -> None:
-        self.tol_edit.copyDataToInputs(new_data.get("tolerancje"))
-        if new_data.get("tol_mode") == "deviations":
-            self.tol_edit.tol_check.setChecked(False)
-            self.tol_edit.dev_check.setChecked(True)
-        else:
-            self.tol_edit.tol_check.setChecked(True)
-            self.tol_edit.dev_check.setChecked(False)
-        self.tol_edit.check.setChecked(new_data.get("use_tol"))
-
-        self.data.material_frame.loadData(new_data.get("material_data"))
-        self.data.zew_dane = new_data.get("zew_dane")
-        self.data.copyDataToInputs(new_data.get("input_dane"))
-
-    def reportData(self) -> str:
-        def indent_point(point_text, bullet, bold, sa=100):
-            bullet_code = "\\bullet" if bullet else ""
-            bold_code = "\\b" if bold else ""
-            return f"{{\\pard\\sa{str(sa)}\\fi-300\\li600{bullet_code}{bold_code}\\tab {point_text} \\par}}"
-
-        if not self.use_this_check.isChecked():
-            return ''
-        materials = self.data.material_frame.getData()
-        wyniki = obliczenia_mech_wyjsciowy(self.data.input_dane, self.data.zew_dane, materials, self.data.tol_data, self.data.wheel_rotation_angle)
-
-        text = "{\\pard\\b Mechanizm wyjściowy \\line\\par}"
-        text += "{\\pard\\sa200\\b Dane: \\par}"
-        text += indent_point(f"Liczba sworzni z tulejami: n = {self.data.input_dane['n']}", True, True)
-        text += indent_point(f"Promień rozmieszczenia sworzni z tulejami: R{{\sub wt}} = {self.data.input_dane['R_wt']} [mm]", True, True)
-        if self.data.zew_dane["K"] == 2:
-            text += indent_point(f"Odstęp pomiędzy kołami: x = {self.data.input_dane['e2']} [mm]", True, True)
-        
-        text += "{\\pard\\sa200\\b Materiały: \\par}"
-        text += "{\\pard\\sa100 koło cykloidalne: \\par}"
-        text += indent_point(f"Materiał: {materials['wheel']['nazwa']}", False, False)
-        text += indent_point(f"Moduł Younga: E = {materials['wheel']['E']} [MPa]", False, False)
-        text += indent_point(f"Liczba Poissona: v = {materials['wheel']['v']}", False, False)
-        text += "{\\pard\\sa100 tuleja: \\par}"
-        text += indent_point(f"Materiał: {materials['sleeve']['nazwa']}", False, False)
-        text += indent_point(f"Moduł Younga: E = {materials['sleeve']['E']} [MPa]", False, False)
-        text += indent_point(f"Liczba Poissona: v = {materials['sleeve']['v']}", False, False)
-        text += "{\\pard\\sa100 sworzeń: \\par}"
-        text += indent_point(f"Materiał: {materials['pin']['nazwa']}", False, False)
-        text += indent_point(f"Granica plastyczności: R{{\sub e}} {materials['pin']['Re']} [MPa]", False, False)
-        text += indent_point(f"Współczynnik bezpieczeństwa: k = {materials['pin_sft_coef']}", False, False)
-        
-        text += f"{{\\pard\\sa100 Współczynnik tarcia tocznego pomiędzy otworami a tulejami: f{{\sub k-t}} = {self.data.input_dane['f_kt']:.5f} [m]\\par}}"
-        text += f"{{\\pard\\sa100 Współczynnik tarcia tocznego pomiędzy tulejami a sworzniami: f{{\sub t-s}} = {self.data.input_dane['f_ts']:.5f} [m]\\par}}"
-        text += indent_point(f"Nacisk dopuszczalny (dla pary materiałów): p{{\sub dop}} = {materials['p_dop']} [MPa]", False, False)
-
-        text += "{\\pard\\sa200\\b Obliczenia: \\par}"
-        text += indent_point("Siły działające na sworzeń:", True, True)
-        text += indent_point(f"Maksymalna siła działająca na sworzeń: F{{\sub max}} = {wyniki['F_max']} [N]", False, False)
-        text += indent_point(f"Wypadkowa siła w mechanizmie: F{{\sub wmr}} = {self.data.obliczone_dane['F_wmr']} [N]", False, False)
-        text += indent_point(f"Ramię działania siły wypadkowej w mechanizmie: r{{\sub mr}} = {self.data.obliczone_dane['r_mr']} [mm]", False, False, 500)
-        
-        text += indent_point("Geometria mechanizmu wyjściowego:", True, True)
-        text += indent_point(f"Sposób podparcia sworznia: {self.data.input_dane['podparcie']}", False, False)
-        text += indent_point(f"Obliczona średnica sworznia: d{{\sub sobl}} = {self.data.obliczone_dane['d_sw']} [mm]", False, False)
-        text += indent_point(f"Przyjęta średnica sworznia: d{{\sub s}} = {self.data.input_dane['d_sw']} [mm]", False, False)
-        text += indent_point(f"Obliczona średnica zewnętrzna tulei: d{{\sub tzobl}} = {self.data.obliczone_dane['d_tul']} [mm]", False, False)
-        text += indent_point(f"Przyjęta średnica zewnętrzna tulei: d{{\sub tz}} = {self.data.input_dane['d_tul']} [mm]", False, False)
-        text += indent_point(f"Średnica otworu pod tuleje: d{{\sub o}} = {self.data.obliczone_dane['d_otw']} [mm]", False, False, 500)
-        
-        text += indent_point("Naciski pomiędzy tulejami a otworami:", True, True)
-        text += indent_point(f"Maksymalne naciski pomiędzy tuleją a otworem: p{{\sub max}} = {wyniki['p_max']} [MPa]", False, False)
-        text += f"{{\\pard\\sa500\\qc Warunek p{{\sub max}} = {wyniki['p_max']} [MPa] < p{{\sub dop}} = {materials['p_dop']} [MPa] został spełniony. \\par}}"
-        
-        text += indent_point("Moc tracona:", True, True)
-        text += indent_point(f"Całkowita strata mocy: N{{\sub Cmr}} = {round(sum(wyniki['straty'][0]), 3)} [W]", False, False, 500)
-
-        a = "{\\trowd\\trleft3200"
-        b = "{\\trowd\\trleft4000"
-        borders = "\\clbrdrt\\brdrw15\\brdrs\\clbrdrl\\brdrw15\\brdrs\\clbrdrb\\brdrw15\\brdrs\\clbrdrr\\brdrw15\\brdrs\\clvertalc"
-        start = "\\pard\\intbl\\qc "
-        end = "\\cell"
-        endrow = "\\row}"
-
-        text += b + borders + "\\cellx4800" + borders + "\\cellx5600" + borders + "\\cellx6400" + start + "F{\sub j}" + end + start + "p{\sub j}" + end + start + "N{\sub mrj}" + end + endrow
-        text += a + borders + "\\cellx4000" + borders + "\\cellx4800" + borders + "\\cellx5600" + borders + "\\cellx6400" + start + "Nr sworznia" + end + start + "[N]" + end + start + "[MPa]" + end + start + "[W]" + end + endrow
-        for index, row in enumerate(zip(wyniki["sily"][0], wyniki["naciski"][0], wyniki["straty"][0]), 1):
-            text += a + borders + "\\cellx4000" + borders + "\\cellx4800" + borders + "\\cellx5600" + borders + "\\cellx6400" + start + str(index) + end + start + "{:.1f}".format(row[0]) + end + start + "{:.2f}".format(row[1]) + end + start + "{:.3f}".format(row[2]) + end + endrow
-
-        text += "\\line"
-        return text
-
-    def csvData(self) -> str:
-        if not self.use_this_check.isChecked():
-            return ''
-        wyniki = obliczenia_mech_wyjsciowy(self.data.input_dane, self.data.zew_dane, self.data.material_frame.getData(), self.data.tol_data, self.data.wheel_rotation_angle)
-        title = "Mechanizm wyjściowy ze sworzniami\n"
-        sily_text = [f"{i},{round(wyniki['sily'][0][i], 1)}\n" for i in range(1, len(wyniki['sily']) + 1)]
-        naciski_text = [f"{i},{round(wyniki['naciski'][0][i], 2)}\n" for i in range(1, len(wyniki['naciski']) + 1)]
-        straty_text = [f"{i},{round(wyniki['straty'][0][i], 3)}\n" for i in range(1, len(wyniki['straty']) + 1)]
-        return title + "Siły na sworzniach [N]\n".join(sily_text) + "Naciski powierzchniowe na sworzniach [MPa]\n".join(naciski_text) + "Straty mocy na sworzniach [W]\n".join(straty_text) + "\n"
