+++ Tabs/pin_out/subtabs/entry_data/popups.py
