+++ Tabs/pin_out/subtabs/entry_data/utils.py
