+++ Tabs/pin_out/subtabs/entry_data/widgets.py
