+++ main_window/view/main_window.py
@@ -11,7 +11,9 @@ from .error_widget import ErrorWidget
 
 from tabs.gear.view.gear_tab import GearTab
 from tabs.gear.controller.gear_tab_controller import GearTabController
-from tabs.pin_out.pin_out_tab import PinOutTab
+
+from tabs.pin_out.view.pin_out_tab import PinOutTab
+from tabs.pin_out.controller.pin_out_tab_controller import PinOutTabController
 from tabs.input_shaft.input_shaft_tab import InputShaftTab
 from tabs.output_mechanism.output_mechanism_tab import OutputMechanismTab
 
@@ -62,9 +64,10 @@ class MainWindow(QMainWindow):
         self.gear_tab_controller = GearTabController(self.gear_tab)
         self.gear_tab.data.animDataUpdated.connect(self.updateAnimationData)
 
-        # self.pin_out_tab = PinOutTab(self.central_widget)
-        # self.pin_out_tab.data.animDataUpdated.connect(self.updateAnimationData)
-        # self.gear_tab.data.dane_materialowe.wheelMatChanged.connect(self.pin_out_tab.data.material_frame.changeWheelMat)
+        self.pin_out_tab = PinOutTab(self.central_widget)
+        self.pin_out_tab_controller = PinOutTabController(self.pin_out_tab)
+        self.pin_out_tab.data.animDataUpdated.connect(self.updateAnimationData)
+        self.gear_tab.data.dane_materialowe.wheelMatChanged.connect(self.pin_out_tab.data.material_frame.changeWheelMat)
 
         # self.output_mechanism_tab = OutputMechanismTab(self.central_widget)
 
@@ -74,9 +77,9 @@ class MainWindow(QMainWindow):
 
         # self.input_shaft_tab = InputShaftTab(self.central_widget)
 
-        self.tab_titles = ["Zarys"] #, "Mechanizm Wyj I", "Mechanizm Wyj II", "Mechanizm Wej"]
-        self.tab_widgets = [self.gear_tab] #, self.pin_out_tab, self.output_mechanism_tab, self.input_shaft_tab]
-        self.tab_controllers = [self.gear_tab_controller]
+        self.tab_titles = ["Zarys", "Mechanizm Wyj I"] #, "Mechanizm Wyj II", "Mechanizm Wej"]
+        self.tab_widgets = [self.gear_tab, self.pin_out_tab] # self.output_mechanism_tab, self.input_shaft_tab]
+        self.tab_controllers = [self.gear_tab_controller, self.pin_out_tab_controller]
 
         for index, (title, tab) in enumerate(zip(self.tab_titles, self.tab_widgets)):
             button = QPushButton(title)
@@ -99,7 +102,7 @@ class MainWindow(QMainWindow):
         self.error_box = ErrorWidget(self.central_widget)
         self.error_box.show()
         self.gear_tab.data.errorsUpdated.connect(partial(self.error_box.updateErrors, module="GearTab"))
-        # self.pin_out_tab.data.errorsUpdated.connect(partial(self.error_box.updateErrors, module="PinOutTab"))
+        self.pin_out_tab.data.errorsUpdated.connect(partial(self.error_box.updateErrors, module="PinOutTab"))
         self.error_box.resetErrors()
 
         # Set base data widget
@@ -222,8 +225,7 @@ class MainWindow(QMainWindow):
             tab_controller.receiveData(passed_data)
     
     def onAnimationTick(self, kat):
-        pass
-        # self.pin_out_tab.data.recalculate(kat)
+        self.pin_out_tab.data.recalculate(kat)
     
     def updateAnimationData(self, dane):
         self.animation_view.animation.updateData(dane)
